Got a basic file loading working for when I start loading shaders.  Created more configurations for different testing reasons.
Realized that the UE style logging I could've just been a #define LOG_MESSAGE() sort of thing.  I guess I was thinking of using a templated function instead would be cool.  I guess I could also do more with verbosity.  But it does what it needs to for what I will be using it for.